Move cursor with explicit direction enum
I'm not sure about this. The idea is to make this more general, so that
other functionality can be implemented in terms of `cursor_move`. This
would take care of the clamping / normalization. However, some insert
mode operations actually place the cursor on after the end of the line.
I haven't decided how I want to handle this yet.

This commit also switches to counting only source lines, not blanks or
comments.